fix: 类型错误修复

+++ PageRequest.go
@@ -1,18 +1,18 @@
 package data_common
 
 type PageRequest struct {
-	Page uint
-	Size uint
+	Page int64
+	Size int64
 }
 
-func (r PageRequest) GetPageSize() uint {
+func (r PageRequest) GetPageSize() int64 {
 	return r.Size
 }
 
-func (r PageRequest) GetPageNumber() uint {
+func (r PageRequest) GetPageNumber() int64 {
 	return r.Page
 }
 
-func (r PageRequest) GetOffset() uint {
+func (r PageRequest) GetOffset() int64 {
 	return r.Page * r.Size
 }
